requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
pandas==2.2.2
numpy<2
scikit-learn==1.5.1
apache-airflow==2.9.3
requests==2.32.3
EOF